requirements.txt
python-dotenv
langchain-community
langchain-google-genai
langchain-core
langchain
faiss-cpu
pypdf
flask
google-generativeai
fastapi
uvicorn
pydantic
pandas